Change to SVG, update task IDs, and show arg
Change png to svg. This results in a better image, but it must be opened
with a web browser. For future state, add an option to specify format.
Also, do not use feh for svg.

Task IDs were not shown for waiting tasks. This makes it harder to edit
them. The code was updated to show all task IDs except those using a
UUID such as completed or deleted tasks.

A new argument was added, -s or --show. This is meant to start the
viewer.