Enable re-run last report if no user args
Added exceptions and redid the return codes so that when no user
arguments are provided the last query is used but Taskwarrior is not
run, otherwise it would just run the default report for the command
`task`.

Also, a conditional was added to correctly output SVG on Linux. This
commit has only been tested on Windows.