Fix bug preventing multiple arguments
The subprocess package requires each argument to be a separate item in
the list. Depending on how the user enters the commands, some items in
the list can contain multiple arguments. For example:

['+READY pro:test_project', '+DELETED']

To convert this properly, first the list must be combined using 'join'
so there is only one string and then split apart using `split`. The
final format is then:

['+READY', 'pro:test_project', '+DELETED']